Fix OffAmazonPayments check; emit login information
'window.OffAmazonPayments' in window returns false even when it is
appropriately loaded, focusing the document.write line to always
execute.

This commit also expands the emit('login') call to include the response
information. This is helpful because response includes the access token,
which one can use to retrieve information about the user's profile.